Added files
discordBot is the main file to run. It pulls its commands from botCmds and uses pre to manage settings and shared variables.
I know I could've made this bot a lot simpler if I used the ext.commands part of the discord API
but where's the fun in that? I did this to challenge myself more than anything.

You'll need to pip install discord.py which can be found here:
https://github.com/Rapptz/discord.py